refactor(mcp): move McpServerSummary out of UI slash types into src/mcp/summary.ts
- New src/mcp/summary.ts: canonical McpServerSummary interface + buildMcpServerSummary() builder
- slash/types.ts: import+re-export McpServerSummary from MCP layer, remove inline interface and unused inspect/registry imports
- chat.tsx: use buildMcpServerSummary() instead of inline object literal

Addresses issue #196 first boundary: relocation + builder, zero behavior change.
